Narrow Amount back from long to int
Reverts the widening in 079caec. int caps a single payment at 2,147,483,647
minor units (~£21.4M), which is well clear of anything this gateway is
expected to carry, and it keeps the wire contract to the conventional type.

Carried through the request, both responses, Payment, the service request and
the bank wire DTO, so the type is consistent across the whole path. The
integration test that asserted an amount past int.MaxValue goes with it.

+++ src/PaymentGateway.Infrastructure/AcquiringBank/AcquiringBankPaymentRequestDto.cs
@@ -6,5 +6,5 @@ internal record AcquiringBankPaymentRequestDto(
     [property: JsonPropertyName("card_number")] string CardNumber,
     [property: JsonPropertyName("expiry_date")] string ExpiryDate,
     [property: JsonPropertyName("currency")] string Currency,
-    [property: JsonPropertyName("amount")] long Amount,
+    [property: JsonPropertyName("amount")] int Amount,
     [property: JsonPropertyName("cvv")] string Cvv);
